packaging: Adding initial script for Windows
Can't say anything special here. The difference between this and the
"old" cura-build is that we are calling now a python script, which is
now using cx_Freeze. The results are then moved like before from Py2Exe
to NSIS and then the installer is built (triggered by CPack).

Contributes to CURA-2150