packaging: Windows: Adding missing icon
At the moment the Cura icon on "Programs and Features" is missing.
This commit puts our icon into your installation again and sets the name
as CPACK_NSIS_INSTALLED_ICON_NAME, which was missing before and thus the
icon was missing in the list of applications in "Programs and Features",
too.

Contributes to CURA-2292