Run Cura without elevation from the installer
When Cura is launched by the installer, it is run "As Administrator" (elevated), because the installer process runs elevated. This causes drag and drop not to work on first run, and this causes python cache files to be written to the program files folder (which may in turn cause other issues).

Launching Cura through explorer.exe ensures that cura is ran as the currently logged in user. For details see http://mdb-blog.blogspot.nl/2013/01/nsis-lunch-program-as-user-from-uac.html