packaging: Adding CLI application
This will add a blocking CLI application. Normally after execution the application will be spawned in the background in Windows (it seems to me). However, since we are interested in headless usage of Cura, we need an executable, which is blocking until it is done. Anything else makes no sense for this use case.