Using the built-in CPACK_PACKAGE_FILE_NAME variable to set the resulting package name of the installer (until now that was done automatically).
For WIX we need to have the CPACK_PACKAGE_NAME variable set to "Ultimaker Cura" as it initially was (i.e., with spaces), since it is being used while creating the uninstaller shortcut and we do not want any underscores in any shortcut. Hence, if we want the final package name of the installer to be without spaces we need to explicitly use a variable that so far was created automatically during build time, namely CPACK_PACKAGE_FILE_NAME.
Another change here is the final value of CPACK_PACKAGE_NAME, which for user experience purposes is set to e.g., "Ultimaker Cura X.Y", so that the eventual uninstaller shortcut will be easily distinguishable. So far, all .msi uninstallers were of the type "Uninstall Ultimaker Cura", which would be overriden in case a newer release is installed. With the new value now it will be "Uninstall Ultimaker Cura X.Y".