'Downgrade' to Python 3.8 for Windows.
Windows 7 (which we might drop in the near future, but for now our build-server is also Win7) can't build Python 3.9 (support was dropped in any case). Downgrade is between quotes because we never actually used 3.9 in a build yet (was only ever a brief thing in this upgrade python branch, when we didn't know about the dropped support yet).

CURA-7900