Reduce optimisation level by 1
This is necessary because the Numpy library requests for the docstring of things when importing, but we remove the docstring function from our builds with this optimisation flag. This then crashes the application when importing numpy. Very annoying. Let's see what sort of impact this'll have.

Contributes to issue CURA-7900.