Use exact Cura python version.
After the latest Python upgrade the version we are currently using is v3.8.8. However, the macOS build VM on AWS had already a system Python version installed by default, which is more recent than ours (v3.9.1).
As a result, during CMake's configuration step the system Python is found first (since it's located in `/usr/bin` or `/usr/local/bin`) and, thus, the step results in error.
Using an exact version of Python is the most ideal solution for this issue.