Move the two files related to signing the app on macOS within the repository
With the latest security enhancements in macOS versions >= 10.15 (Catalina) all the libraries within the application bundle have to be signed.
In our case that means all .so and .dylib files and then the whole .app needs to be signed as well.
CMake commands do not work very well with multiple, piped shell commands so we created a dedicated script to handle that dor us.
The script itself references the cura.entitlements file which is an extra file that basically grants the final executable with certain capabilities.
With this commit we add the two helper files in the repository, so that they're available during the signing stage of Cura.