Add CURA_VERSION_EXTRA to the package file name
Until now, we've been manually renaming the installer on the build system after its creation. However, sometimes this task fails because of ssh connection error to the build server, thus making the whole build job to fail, even though the installer has been successfully built.
As a result, it's better to add the extra version to the package file name from the start in order to avoid ssh-ing to the build server too many times during a build job.