Include only CURA_VERSION to the installer name
The CURA_VERSION variable contains all that we need, i.e., the major, minor, patch and extra version (if applicable) of Cura.
So far, we've been adding the extra version to the installer's name on the build system after connecting again to the build server. Lately, however, there have been some cases where the ssh connection to perform this change would fail, thus making the whole build job fail as well. As a result, it is much better to have the extra version added to the file name via CPack from the start and avoid such ssh connection errors.