Add libraries that scipy.spatial needs while initializing
After using scipy v1.6.3 on Linux, the AppImage cannot launch due to a `File Not Found` error. It seems that scipy.spatial requires the `libopenblasp-r0-085ca80a.3.9.so` and `libgfortran-ed201abd.so.3.0.0.so` during initialization, so we are adding them to the list of files to be included in the final AppImage.