Remove bin_excludes
We had these four binaries to be excluded, because `cx-Freeze` was trying to add them from the `.dylibs` folder of both `scipy` and `shapely` and it would then crash.
Since this is not required anymore, we do not exclude them, as that would basically revert the stage where we add them from numpy. 
Because of that there was no `libquadmath.0.dylib` present in the parent folder of the application bundle, which would force Cura to crash on Big Sur and Monterey, since the linker was expecting this file to be part of the system path of gcc v10, but the gcc versions of those two macOS installations use higher gcc versions.