Update list of DLLs to include into build
Added TextToSpeech which was missing. We don't seem to use it now, but it gets included on other platforms so this is consistent.
Removed NetworkAuth which doesn't seem to exist any more.
Commented out a bunch of plug-ins that we don't build for other platforms, to make the installation lighter and get more consistent behaviour between platforms.

Contributes to issue CURA-8592.