Re-uploading the .ico file for Windows
Apparently the .ico file cannot be larger than 256x256, so the creation of the installer is failing when using the previous one uploaded.
This is an attempt to upload the icon being 256x256.
For some reason though the icon looks a bit blurry and pixelated, so we would to investigate why this is.