Update Qt library paths to look in Python site-packages
Because we install Qt6 via the Pip package, it now gets placed in site-packages in the PyQt6 folder, instead of in the main lib folder. We copy the QML libraries from it so we need to adjust the path.
I also updated the libraries we use, because some of those changed:
- qml/Qt disappeared. Not sure what happened with it though.
- qml/QtQuick.2 got renamed to qml/QtQuick, and the original QtQuick 1 is removed.
- qml/QtGraphicalEffects got removed; it is no longer supported by Qt6.
- I added qml/QtPositioning, since it is a new library necessary for the new TableView implementation.

Contributes to issue CURA-7924.